+++ src/main/java/com/tribes_backend/tribes/userModel/UserModelHelpersMethods.java
@@ -24,16 +24,5 @@ public class UserModelHelpersMethods {
         }
         return false;
     }
-    public String getErrorParameter (TribesUser user){
-        String toReturn = "";
-        if (user.getUsername().isEmpty() || user.getUsername() == null){
-            toReturn = "username";
-            return toReturn;
-        }
-        else if (user.getPassword().isEmpty() || user.getPassword() == null){
-            toReturn = "password";
-            return toReturn;
-        }
-        return null;
-    }
+
 }
