+++ src/main/resources/db/migration/V1__test_init.sql
@@ -1,2 +1,1 @@
-CREATE TABLE USERS (ID INT AUTO_INCREMENT PRIMARY KEY, USERID VARCHAR(45));
-INSERT INTO USERS (ID, USERID) VALUES (1, 'tutorialspoint.com');
+CREATE TABLE USERS (ID INT AUTO_INCREMENT PRIMARY KEY, USERNAME VARCHAR(45), PASSWORD VARCHAR(90));
